separate Card, List components into `features` directory
