+++ client/src/components/features/index.ts
@@ -1,5 +1,7 @@
-export { default as AddItemForm } from './AddItemForm'
-export { default as CardDraggable } from './CardDraggable'
-export { default as CardDroppable } from './CardDroppable'
-export { default as ListDraggable } from './ListDraggable'
-export { default as ListDroppable } from './ListDroppable'
+export { default as AddItemForm } from './AddItemForm';
+export { default as CardDraggable } from './CardDraggable';
+export { default as CardDroppable } from './CardDroppable';
+export { default as ListDraggable } from './ListDraggable';
+export { default as ListDroppable } from './ListDroppable';
+export { default as Card } from './Card';
+export { default as List } from './List';
