+++ client/src/pages/index.ts
@@ -1,3 +1,1 @@
-export { default as Board } from './board/Board';
-export { default as BoardList } from './board/BoardList';
-export { default as ListCard } from './board/ListCard';
+export { default as Board } from './Board';
